srcs/backend/requirements.txt
Django
gunicorn
psycopg2-binary
django-cors-headers
djangorestframework
requests
djangorestframework-simplejwt
channels
channels_redis
aiohttp
uvicorn
whitenoise
django-extensions
daphne
pyjwt
qrcode
pyotp